requirements.txt
python-dotenv
langchain-google-genai
langchain
langchain-community
pdfplumber
sentence-transformers
chromadb
fastapi
uvicorn
